MCP/requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
requests>=2.31.0
aiohttp>=3.9.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
motor==3.3.2
pymongo==4.6.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0 